split out myft buttons stuff so the index is just the index
