+++ myft/ui/personalise-links.js
@@ -1,0 +1,10 @@
+const myFtClient = require('next-myft-client');
+const $$ = require('n-ui-foundations').$$
+
+export default function (el) {
+	const links = (el && el.nodeName === 'A') ? [el] : $$('a[href^="/myft"]', el);
+	links.forEach(link => {
+		myFtClient.personaliseUrl(link.getAttribute('href'))
+			.then(personalisedUrl => link.setAttribute('href', personalisedUrl));
+	});
+}
